Refactor: merge Markdown{News,Server}Loader into loaders.py
